Allow compression types to be optional.
