+++ plugins/sqlite/io/SQLiteReader.cpp
@@ -33,9 +33,11 @@
 ****************************************************************************/
 
 #include "SQLiteReader.hpp"
+
 #include <pdal/PointView.hpp>
 #include <pdal/pdal_macros.hpp>
 #include <pdal/PDALUtils.hpp>
+#include <pdal/compression/LazPerfCompression.hpp>
 #include <pdal/util/ProgramArgs.hpp>
 
 namespace pdal
